requirements/prod.txt
django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.11
requests==2.32.5
Pillow==12.0.0
pymongo==4.15.5
dnspython==2.8.0
gunicorn==21.2.0
